chore(refactor): separate utils file and use dynamic lookup for sequencers in smoke test
